Rename llm_inputs directory and file to inputs
