+++ genai-perf/docs/files.md
@@ -69,7 +69,7 @@ versus the input sequence length for each request.
 
 ##### JSON Files
 
-- llm_inputs.json: This contains the input prompts provided to the LLM during testing.
+- inputs.json: This contains the input prompts provided to the LLM during testing.
 - profile_export.json: This is provided by Perf Analyzer and contains the timestamps
 for each event in the lifecycle of each request. This is low-level data used to calculate
 metrics by GenAi-Perf.
