+++ genai-perf/genai_perf/inputs/farewell.txt
