+++ genai-perf/genai_perf/inputs/source_images/h100.jpeg
