+++ genai-perf/genai_perf/inputs/source_images/h200.jpeg
