+++ genai-perf/genai_perf/inputs/source_images/jensen.jpeg
