+++ genai-perf/genai_perf/inputs/synthetic_image_generator.py
