+++ genai-perf/genai_perf/inputs/synthetic_prompt_generator.py
