+++ genai-perf/genai_perf/parser.py
@@ -41,13 +41,13 @@ from genai_perf.constants import (
     DEFAULT_TRITON_METRICS_URL,
     OPEN_ORCA,
 )
-from genai_perf.llm_inputs.llm_inputs import (
+from genai_perf.inputs.inputs import (
     LlmInputs,
     ModelSelectionStrategy,
     OutputFormat,
     PromptSource,
 )
-from genai_perf.llm_inputs.synthetic_image_generator import ImageFormat
+from genai_perf.inputs.synthetic_image_generator import ImageFormat
 from genai_perf.plots.plot_config_parser import PlotConfigParser
 from genai_perf.plots.plot_manager import PlotManager
 from genai_perf.tokenizer import DEFAULT_TOKENIZER
